Carry over hash value for all literal types in remote caching
Previously, the hash set on an input literal would be ignored if
the literal was of either the `Collection` or `Map` type. This is
fixed here by moving the hash-checking branch up in front of the
special type logic, so that an empty literal with hash is constructed
in either case if a hash is found.

Adds a regression test to verify the hashes are carried over into the
representer literals for both map and collection types.

Signed-off-by: Nicholas Junge <nicholas.junge@web.de>